feat: 좋아요 버튼 컴포넌트로 분리 후 적용 (#29)
