+++ src/components/atoms/button/index.ts
@@ -1,2 +1,3 @@
 export { default as Button } from './Button';
+export { default as LikeButton } from './like/LikeButton';
 export { default as SegmentedButton } from './segmented/SegmentedButton';
