+++ src/components/composites/book/detail/record/list/item/BookRecordListItem.styled.ts
@@ -24,15 +24,3 @@ export const ratingIcon = (isFilled: boolean) => css`
 export const RecordItemContent = tw.p`m-body-r14 tablet:t-body-r14 desktop:d-body-r16`;
 
 export const RecordItemFooter = tw.footer`flex justify-end w-full`;
-
-export const LikeButton = tw.button`flex items-center gap-[4px]`;
-
-export const likeIcon = css`
-  ${tw`w-[16px] h-[16px]`}
-
-  & > g > path {
-    ${tw`fill-blue200`}
-  }
-`;
-
-export const Like = tw.span`m-body-m13 tablet:t-body-m13 desktop:d-body-m15`;
