+++ src/components/composites/book/detail/record/list/item/BookRecordListItem.tsx
@@ -1,16 +1,6 @@
 import React from 'react';
-import { useParams } from 'react-router-dom';
 
-import { useUser } from '@/contexts';
-import { Profile } from '@/components';
-import {
-  useCreateLikeForRecord,
-  useDeleteLikeForRecord,
-  useGetTotalLikeForRecord,
-} from '@/services';
-import { getFirstIsbnSegment } from '@/utils';
-import LikeIcon from '@/assets/icon/ic_thumb_up.svg?react';
-import LikeFilledIcon from '@/assets/icon/ic_thumb_up_filled.svg?react';
+import { LikeButton, Profile } from '@/components';
 import RatingIcon from '@/assets/icon/ic_rating.svg?react';
 import type { GetBookUserRecordsServerModel } from '@/types';
 import * as S from './BookRecordListItem.styled';
@@ -20,23 +10,6 @@ interface BookRecordListItemProps {
 }
 
 const BookRecordListItem = ({ record }: BookRecordListItemProps) => {
-  const { id } = useParams();
-
-  const { user } = useUser();
-  const isbn = getFirstIsbnSegment(id);
-  const userId = user?.id!;
-  const req = { isbn, userId, recordId: record.id };
-
-  const { data } = useGetTotalLikeForRecord(req);
-  const { mutate: createLikeRecord } = useCreateLikeForRecord();
-  const { mutate: deleteLikeRecord } = useDeleteLikeForRecord();
-
-  const handleLikeButtonClick = (recordId: string) => () => {
-    const req = { recordId, userId, isbn };
-
-    data[0].isliked ? deleteLikeRecord(req) : createLikeRecord(req);
-  };
-
   return (
     <S.RecordItemContainer key={record.id}>
       <S.RecordItemHeader>
@@ -53,18 +26,11 @@ const BookRecordListItem = ({ record }: BookRecordListItemProps) => {
       </S.RecordItemHeader>
       <S.RecordItemContent>{record.record_comment}</S.RecordItemContent>
       <S.RecordItemFooter>
-        <S.LikeButton
-          disabled={userId === record.user_id}
-          type="button"
-          onClick={handleLikeButtonClick(record.id)}
-        >
-          {data[0].isliked ? (
-            <LikeFilledIcon css={S.likeIcon} />
-          ) : (
-            <LikeIcon css={S.likeIcon} />
-          )}
-          <S.Like>{data[0].count}</S.Like>
-        </S.LikeButton>
+        <LikeButton
+          isbn={record.isbn}
+          recordId={record.id}
+          userId={record.user_id}
+        />
       </S.RecordItemFooter>
     </S.RecordItemContainer>
   );
